refactor: ChoicePanel UI 제거와 OutputPanel 중심 구조 정리
