--- src/features/choice-panel/ui/index.ts
@@ -1,4 +1,0 @@
-export * from "./ActionStep";
-export * from "./ChoicePanel";
-export * from "./FollowUpStep";
-export * from "./ProcessingMethodStep";
